+++ src/Table.tsx
@@ -149,7 +149,7 @@ export interface TableProps<Row, Key> extends Omit<StandardProps, 'onScroll'> {
   isTree?: boolean;
 
   /** Set the height of an expandable area */
-  rowExpandedHeight?: number;
+  rowExpandedHeight?: ((rowData?: Row) => number) | number;
 
   /** Add an optional extra class name to row */
   rowClassName?: string | ((rowData: Row, rowIndex: number) => string);
@@ -416,7 +416,9 @@ const Table = React.forwardRef(<Row extends RowDataType, Key>(props: TableProps<
     setScrollX,
     getTableHeight
   } = useTableDimension({
-    data: dataProp,
+    // The data should be flattened,
+    // otherwise the array length required to calculate the scroll height in the TreeTable is not real.
+    data,
     width: widthProp,
     rowHeight,
     tableRef,
@@ -473,7 +475,8 @@ const Table = React.forwardRef(<Row extends RowDataType, Key>(props: TableProps<
     onScrollBody,
     onScrollTop,
     onScrollLeft,
-    onScrollTo
+    onScrollTo,
+    onScrollByKeydown
   } = useScrollListener({
     rtl,
     data: dataProp,
@@ -579,7 +582,13 @@ const Table = React.forwardRef(<Row extends RowDataType, Key>(props: TableProps<
 
   const renderRowExpanded = useCallback(
     (rowData?: Row) => {
-      const styles = { height: rowExpandedHeight };
+      let height = 0;
+      if (typeof rowExpandedHeight === 'function') {
+        height = rowExpandedHeight(rowData);
+      } else {
+        height = rowExpandedHeight;
+      }
+      const styles = { height };
 
       if (typeof renderRowExpandedProp === 'function') {
         return (
@@ -938,7 +947,14 @@ const Table = React.forwardRef(<Row extends RowDataType, Key>(props: TableProps<
     if (data) {
       let top = 0; // Row position
       let minTop = Math.abs(scrollY.current);
-      let maxTop = minTop + height + rowExpandedHeight;
+      let startHeight = 0;
+      if (typeof rowExpandedHeight === 'function') {
+        startHeight = data.length ? rowExpandedHeight(data[0]) : 100;
+      } else {
+        startHeight = rowExpandedHeight;
+      }
+
+      let maxTop = minTop + height + startHeight;
       const isCustomRowHeight = typeof rowHeight === 'function';
       const isUncertainHeight = !!renderRowExpandedProp || isCustomRowHeight || wordWrap;
 
@@ -972,7 +988,11 @@ const Table = React.forwardRef(<Row extends RowDataType, Key>(props: TableProps<
             cellHeight = nextRowHeight;
             if (expandedRow) {
               // If the row is expanded, the height of the expanded row is added.
-              nextRowHeight += rowExpandedHeight;
+              if (typeof rowExpandedHeight === 'function') {
+                nextRowHeight += rowExpandedHeight(rowData);
+              } else {
+                nextRowHeight += rowExpandedHeight;
+              }
             }
           }
 
@@ -1107,6 +1127,8 @@ const Table = React.forwardRef(<Row extends RowDataType, Key>(props: TableProps<
         className={classes}
         style={styles}
         ref={tableRef}
+        tabIndex={-1}
+        onKeyDown={onScrollByKeydown}
       >
         {showHeader && renderTableHeader(headerCells, rowWidth)}
         {children && renderTableBody(bodyCells, rowWidth)}
@@ -1152,7 +1174,7 @@ Table.propTypes = {
   renderTreeToggle: PropTypes.func,
   renderRowExpanded: PropTypes.func,
   renderRow: PropTypes.func,
-  rowExpandedHeight: PropTypes.number,
+  rowExpandedHeight: PropTypes.oneOfType([PropTypes.func, PropTypes.number]),
   renderEmpty: PropTypes.func,
   renderLoading: PropTypes.func,
   rowClassName: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),